fix: some changes
