+++ src/components/SwapWidget/MarketOrder/index.tsx
@@ -11,9 +11,9 @@ import { SwapTypes } from 'src/constants/swap';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { MixPanelEvents, useMixpanel } from 'src/hooks/mixpanel';
 import { useTokenHook } from 'src/hooks/tokens';
-import { useCurrency } from 'src/hooks/tokens/evm';
 import { useApproveCallbackFromTradeHook } from 'src/hooks/useApproveCallback';
 import { ApprovalState } from 'src/hooks/useApproveCallback/constant';
+import { useCurrency } from 'src/hooks/useCurrency';
 import useENS from 'src/hooks/useENS';
 import { useSwapCallbackHook } from 'src/hooks/useSwapCallback';
 import useToggledVersion, { Version } from 'src/hooks/useToggledVersion';
