+++ src/components/SwapWidget/SelectTokenDrawer/index.tsx
@@ -6,7 +6,7 @@ import AutoSizer from 'react-virtualized-auto-sizer';
 import { FixedSizeGrid } from 'react-window';
 import Drawer from 'src/components/Drawer';
 import { useChainId } from 'src/hooks';
-import { useToken } from 'src/hooks/tokens/evm';
+import { useTokenHook } from 'src/hooks/tokens';
 import { useAllTokens } from 'src/hooks/useAllTokens';
 import usePrevious from 'src/hooks/usePrevious';
 import { useSelectedListInfo } from 'src/state/plists/hooks';
@@ -39,11 +39,12 @@ const currencyKey = (columnIndex: number, rowIndex: number, data: Currency[], ch
 
 const SelectTokenDrawer: React.FC<Props> = (props) => {
   const { isOpen, onClose, onCurrencySelect, otherSelectedCurrency, selectedCurrency } = props;
+  const chainId = useChainId();
   const [searchQuery, setSearchQuery] = useState<string>('');
   const [isTokenListOpen, setIsTokenListOpen] = useState<boolean>(false);
   const [invertSearchOrder] = useState<boolean>(false);
   const { t } = useTranslation();
-
+  const useToken = useTokenHook[chainId];
   const inputRef = useRef<HTMLInputElement>(null);
   const lastOpen = usePrevious(isOpen);
 
@@ -67,8 +68,6 @@ const SelectTokenDrawer: React.FC<Props> = (props) => {
 
   const allTokens = useAllTokens();
   const selectedListInfo = useSelectedListInfo();
-
-  const chainId = useChainId();
 
   const isAddressSearch = isAddress(searchQuery);
   const searchToken = useToken(searchQuery);
