+++ src/state/pswap/hooks.ts
@@ -21,8 +21,8 @@ import { ROUTER_ADDRESS, ROUTER_DAAS_ADDRESS } from 'src/constants/address';
 import { SWAP_DEFAULT_CURRENCY } from 'src/constants/swap';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { useTradeExactIn, useTradeExactOut } from 'src/hooks/Trades';
-import { useCurrency } from 'src/hooks/tokens/evm';
 import { useHederaTokenAssociated } from 'src/hooks/tokens/hedera';
+import { useCurrency } from 'src/hooks/useCurrency';
 import useParsedQueryString from 'src/hooks/useParsedQueryString';
 import useToggledVersion, { Version } from 'src/hooks/useToggledVersion';
 import { AppState, useDispatch, useSelector } from 'src/state';
